Change the tests from using EhCacheRegionFactory to SingletonEhCacheRegionFactory
The tests rely on there being exactly one unnamed cache. Earlier versions of
EhCache are more tolerant and will reuse an existing unnamed cache if an
attempt is made to create another unnamed cache (or an attempt is made to
create another named cache with the same name), but later versions will throw an
exception. By using SingletonEhCacheProvider, the same cache is always used,
which is what earlier versions of EhCache do anyways, and prevents problems with
later versions.